fix : cover_letter 앱 삭제 및 accounts 로 이동
