--- backend/cover_letter/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class CoverLetterConfig(AppConfig):
-    default_auto_field = "django.db.models.BigAutoField"
-    name = "cover_letter"
