--- backend/cover_letter/views.py
@@ -1,15 +1,0 @@
-# Create your views here.
-def create(request):
-    pass
-
-
-def read(request, id):
-    pass
-
-
-def update(request, id):
-    pass
-
-
-def delete(request, id):
-    pass
